community_server/requirements.txt
fastapi[standard]==0.115.2
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.1
python-dotenv==1.0.0
cryptography==41.0.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
jinja2==3.1.2
requests
auth
python-jose[cryptography]
fastapi_mail
aiosmtplib
dotenv